feat : Add confirm_disassembly_raids and disassembly_text variables to DetailRaidStorage.vue

+++ src/views/DetailRaidStorage/DetailRaidStorage.vue
@@ -35,6 +35,9 @@ const needNewDisk = ref(false);
 const isReady = ref(false);
 const showAddingDiskButton = ref(true);
 
+const confirm_disassembly_raids = ref(false);
+const disassembly_text = ref("Disable");
+
 // const expansionMinDiskSize = ref(0);
 const isLoadingDiskInfoByStorageSpace = ref<boolean>(false);
 const loadRaid = async () => {
@@ -74,7 +77,13 @@ initEstablishRAID();
 import { useRouter } from "vue-router";
 const router = useRouter();
 const isLoadingDisabledButton = ref<boolean>(false);
-const disabledRaid = async (): Promise<void> => {
+async function disabledRaid (): Promise<void> {
+    if (!confirm_disassembly_raids.value) {
+        confirm_disassembly_raids.value = false;
+        disassembly_text.value =  "Confirm";
+        return;
+    }
+    confirm_disassembly_raids.value = true;
     isLoadingDisabledButton.value = true;
     await raid
         .deleteRaid(storagePath ?? "")
@@ -92,6 +101,10 @@ const disabledRaid = async (): Promise<void> => {
         .finally(() => {
             isLoadingDisabledButton.value = false;
         });
+};
+function deselectDisassemblyRaid(): void {
+    confirm_disassembly_raids.value = false;
+    disassembly_text.value = "Disable";
 };
 
 // ejct disk
@@ -318,7 +331,7 @@ const extenedCapacity = (): void => {
                     {{ $t("Format and Disable") }}
                 </span>
 
-                <Button :label="$t('Disable')" severity="accent" size="medium" @click="disabledRaid"
+                <Button :label="$t(disassembly_text)" severity="accent" size="medium" @click="disabledRaid" @focusout="deselectDisassemblyRaid"
                     :loading="isLoadingDisabledButton"></Button>
             </div>
         </div>
